kvmtool: builtin-setup: fix embedded ELF binaries size
lkvm embeds 2 ELF binaries (guest/pre_init and guest/init) which are
written to the disk during the setup of new virtual machines. The
binaries size is gathered thanks to a symbol generated by ld during the
link step (_binary_%s_size). Unfortunately, the code generated by the
compiler doesn't always match our expectation because the symbol is
relocated.

This commit computes the binaries size in a more reliable way thanks to
the subtraction of the end and start addresses of the binaries.